+++ frontend/storybook/src/components/entry/homepage/footer/footer-jigs.ts
@@ -4,18 +4,11 @@ import "@elements/entry/home/TOSORT/column-list";
 export default {
     title: 'Homepage Paragraph',
   }
-  const STR_TITLE="JIGs";
-  const STR_LINE1="Hebrew JIGs";
-  const STR_LINE2="Jewish holidays JIGs";
-  const STR_LINE3="Torah JIGs";
-  const STR_LINE4="J-Stream";
-  const STR_LINE5="Israel";
-  const STR_LINE6="Songs";
-  const STR_WHITE ="white";
+
 
 export const footerjigs= () => {
     return `
-    <footer-jigs head_title="${STR_TITLE}">
+    <footer-jigs>
     
      </footer-jigs>
     `
