+++ frontend/storybook/src/components/entry/homepage/popups/logout-paragraph.ts
@@ -1,5 +1,5 @@
 import "@elements/entry/home/TOSORT/column-list";
-import "@elements/entry/popup/sections/logout-section";
+import "@elements/entry/home/sections/logout-section";
 import "@elements/core/titles/variants/title-section";
  import "@elements/core/buttons/rectangle" ;
  import "@elements/core/popups/popup-container";
